Delete duplicate icon in notifications.html
Today I was working on this element and I discovered that the tests actually failed. This was because the icon name `rv-hookup` was declared twice: once in places and once in notifications. Looking at https://material.io/icons/#ic_rv_hookup, the places icon is correct, but the `notification:rv-hookup` is incorrect. As such, remove the `rv-hookup` in notification.html and make the test pass again.